Updated preseeds to support full-disk encryption
This was not easy as preseeding encrypted volumes is poorly documented.
In the end to save install time we opted to bypass the disk overwrite
process before setting up encrypted volumes which required a live
monkeypatch to partman!

The first phase bootstrap preseed skips worrying about encryption
although it does set up partitions in place for phase 2. The preseed.cfg
file is where the magic is and it will automatically set up full disk
encryption for the user while prompting them for their passphrase. Due
to how Debian crypto preseed works, it required the encrypted volumes to
be inside LVM.

Other fixes include taking advantage of the fact that the new PureOS DVD
includes a functional initrd we can point to instead of downloading one
during bootstrap.